Add CookieService class
