+++ src/main/java/com/epam/tishkin/controller/servlet/LibraryActionsServlet.java
@@ -1,5 +1,6 @@
 package com.epam.tishkin.controller.servlet;
 
+import com.epam.tishkin.controller.CookieService;
 import com.epam.tishkin.controller.TokenManager;
 import com.epam.tishkin.controller.actions.Action;
 import com.epam.tishkin.controller.actions.ActionFactory;
@@ -15,7 +16,7 @@ import java.io.IOException;
 
 @WebServlet("/controller")
 public class LibraryActionsServlet extends HttpServlet {
-    private final TokenManager tokenManager = new TokenManager();
+    private final CookieService cookieService = new CookieService();
     private final static Logger logger = LogManager.getLogger(LibraryActionsServlet.class);
 
     @Override
@@ -30,7 +31,7 @@ public class LibraryActionsServlet extends HttpServlet {
 
     protected void service(HttpServletRequest request, HttpServletResponse response) {
         try {
-            if (tokenManager.verifyToken(request)) {
+            if (cookieService.verifyTokenFromCookie(request)) {
                 Action action = ActionFactory.getAction(request);
                 String view = action.execute(request);
                 request.getRequestDispatcher(view).forward(request, response);
