+++ src/main/webapp/index.jsp
@@ -1,10 +1,10 @@
 <%@page contentType="text/html" pageEncoding="UTF-8"%>
 <html>
-<head>
-<title>Index</title>
-</head>
-<body>
-<jsp:forward page="/jsp/login.jsp"/>
-</body>
+  <head>
+    <title>Index</title>
+  </head>
+  <body>
+    <jsp:forward page="/jsp/login.jsp"/>
+  </body>
 </html>
